+++ src/jackdaw/serdes/avro.clj
@@ -101,9 +101,9 @@
   [^Schema schema]
   (when schema
     (let [base-type (-> schema (.getType) (.getName))
-          logical-type (-> schema (.getProps) (.get "logicalType"))]
+          logical-type (-> schema (.getObjectProps) (.get "logicalType") )]
       (if logical-type
-        {:type base-type :logical-type logical-type}
+        {:type base-type :logical-type (str logical-type)}
         {:type base-type}))))
 
 (defn make-coercion-stack
@@ -394,7 +394,7 @@
       (and (every? (fn [[field-key [^Schema$Field field field-coercion]]]
                      (let [field-value (get clj-map field-key ::missing)]
                        (if (= field-value ::missing)
-                         (.defaultValue field)
+                         (.defaultVal field)
                          (match-clj? field-coercion field-value))))
                    field->schema+coercion)
            (empty? unknown-fields))))